Removed overhead during plugin activation
