+++ GroupMeClient.Core/Startup.cs
@@ -5,6 +5,7 @@ using GroupMeClient.Core.Services;
 using GroupMeClient.Core.Settings;
 using GroupMeClient.Core.Tasks;
 using GroupMeClient.Core.ViewModels;
+using GroupMeClientPlugin.GroupChat;
 
 namespace GroupMeClient.Core
 {
@@ -25,6 +26,7 @@ namespace GroupMeClient.Core
             SimpleIoc.Default.Register(() => new PersistManager(startupParameters.PersistFilePath));
             SimpleIoc.Default.Register(() => new SettingsManager(startupParameters.SettingsFilePath));
             SimpleIoc.Default.Register(() => new PluginInstaller(startupParameters.PluginPath));
+            SimpleIoc.Default.Register<PluginHost>();
         }
 
         /// <summary>
@@ -36,6 +38,11 @@ namespace GroupMeClient.Core
             SimpleIoc.Default.Register<SearchViewModel>(createInstanceImmediately: false);
             SimpleIoc.Default.Register<StarsViewModel>(createInstanceImmediately: false);
             SimpleIoc.Default.Register<SettingsViewModel>(createInstanceImmediately: false);
+
+            // UI integration is provided via the Seach page for the show-in-context feature.
+            SimpleIoc.Default.Register<IPluginUIIntegration>(
+                () => SimpleIoc.Default.GetInstance<SearchViewModel>(),
+                createInstanceImmediately: false);
         }
 
         /// <summary>
