+++ GroupMeClient.WpfUI/ViewModels/MainViewModel.cs
@@ -254,7 +254,6 @@ namespace GroupMeClient.WpfUI.ViewModels
             // Setup messaging
             Messenger.Default.Register<Core.Messaging.UnreadRequestMessage>(this, this.UpdateNotificationCount);
             Messenger.Default.Register<Core.Messaging.DisconnectedRequestMessage>(this, this.UpdateDisconnectedComponentsCount);
-            Messenger.Default.Register<Core.Messaging.RunPluginRequestMessage>(this, this.IndexAndRunCommand);
             Messenger.Default.Register<Core.Messaging.SwitchToPageRequestMessage>(this, this.SwitchToPageCommand);
             Messenger.Default.Register<Core.Messaging.RebootRequestMessage>(this, (r) => this.RebootReasons.Add(r.Reason), true);
             Messenger.Default.Register<Core.Messaging.DialogRequestMessage>(this, this.OpenBigPopup);
@@ -506,11 +505,6 @@ namespace GroupMeClient.WpfUI.ViewModels
             {
                 this.IsReconnecting = this.DisconnectedComponentCount > 0 || this.TaskManager.RunningTasks.Count > 0;
             });
-        }
-
-        private void IndexAndRunCommand(Core.Messaging.RunPluginRequestMessage cmd)
-        {
-            this.SearchViewModel.RunPlugin(cmd.MessageContainer, cmd.Plugin);
         }
 
         private void SwitchToPageCommand(Core.Messaging.SwitchToPageRequestMessage cmd)
